Put custom invoices routes in one file
